doWrite 작성자 정보 저장, doModify, doDelete 권한체크

+++ src/main/java/com/koreaIT/demo/controller/UsrArticleController.java
@@ -11,6 +11,8 @@ import com.koreaIT.demo.util.Util;
 import com.koreaIT.demo.vo.Article;
 import com.koreaIT.demo.vo.ResultData;
 
+import jakarta.servlet.http.HttpSession;
+
 @Controller
 public class UsrArticleController {
 	
@@ -22,16 +24,21 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
-	public ResultData <Article>  doWrite(String title, String body) {
-
+	public ResultData<Article> doWrite(HttpSession session, String title, String body) {
+		
+		if (session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-L", "로그인 후 이용해주세요");
+		}
+		
 		if (Util.empty(title)) {
 			return ResultData.from("F-1", "제목을 입력해주세요");
 		}
-
+		
 		if (Util.empty(body)) {
 			return ResultData.from("F-2", "내용을 입력해주세요");
 		}
-		articleService.writeArticle(title, body);
+		
+		articleService.writeArticle((int) session.getAttribute("loginedMemberId"), title, body);
 		
 		int id = articleService.getLastInsertId();
 		
@@ -40,25 +47,25 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/showList")
 	@ResponseBody
-	public ResultData <List<Article>> showList() {
-
+	public ResultData<List<Article>> showList() {
+		
 		List<Article> articles = articleService.getArticles();
-
+		
 		if (articles.size() == 0) {
 			return ResultData.from("F-1", "게시물이 존재하지 않습니다");
 		}
-
+		
 		return ResultData.from("S-1", "게시물 목록", articles);
 	}
 	
 	@RequestMapping("/usr/article/showDetail")
 	@ResponseBody
-	public ResultData <Article> showDetail(int id) {
+	public ResultData<Article> showDetail(int id) {
 		
 		Article article = articleService.getArticleById(id);
 		
 		if (article == null) {
-			return ResultData.from("F-1",Util.f("%d번 게시물은 존재하지 않습니다", id));
+			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
 		}
 		
 		return ResultData.from("S-1", Util.f("%d번 게시물", id), article);
@@ -66,12 +73,16 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
+	public ResultData doModify(HttpSession session, int id, String title, String body) {
 		
 		Article article = articleService.getArticleById(id);
 		
 		if (article == null) {
 			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
+		}
+		
+		if ((int) session.getAttribute("loginedMemberId") != article.getMemberId()) {
+			return ResultData.from("F-A", "해당 게시물에 대한 권한이 없습니다");
 		}
 		
 		articleService.modifyArticle(id, title, body);
@@ -81,7 +92,7 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpSession session, int id) {
 		
 		Article article = articleService.getArticleById(id);
 		
@@ -89,10 +100,13 @@ public class UsrArticleController {
 			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
 		}
 		
+		if ((int) session.getAttribute("loginedMemberId") != article.getMemberId()) {
+			return ResultData.from("F-A", "해당 게시물에 대한 권한이 없습니다");
+		}
+		
 		articleService.deleteArticle(id);
 		
 		return ResultData.from("S-1", Util.f("%d번 게시물을 삭제했습니다", id));
 	}
 	
-
 }
